pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "alan910127", email = "alan469357@gmail.com"},
]
dependencies = [
    "fastapi>=0.87.0",
    "uvicorn>=0.20.0",
    "prisma>=0.7.1",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.5",
]
requires-python = ">=3.10"
license = {text = "MIT"}
[project.optional-dependencies]

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool]
[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "mypy>=0.991",
]

[tool.pdm.scripts]
dev = "uvicorn src:app --reload"
